relocated Wiki service files
